+++ cmd/rabtap/message_writer.go
@@ -1,3 +1,4 @@
+// write messages to writers and files
 // Copyright (C) 2017-2019 Jan Delgado
 
 package main
@@ -41,6 +42,7 @@ type RabtapPersistentMessage struct {
 	// rabtap specific fields
 	XRabtapReceivedTimestamp time.Time
 
+	// will be serialized as base64
 	Body []byte
 }
 
@@ -54,7 +56,6 @@ func CreateTimestampFilename(t time.Time) string {
 // NewRabtapPersistentMessage creates RabtapPersistentMessage object
 // from a rabtap.TapMessage
 func NewRabtapPersistentMessage(message rabtap.TapMessage) RabtapPersistentMessage {
-
 	m := message.AmqpMessage
 	return RabtapPersistentMessage{
 		Headers:                  m.Headers,
@@ -95,14 +96,27 @@ func (s RabtapPersistentMessage) ToAmqpPublishing() amqp.Publishing {
 		Body:            s.Body}
 }
 
-// WriteMessageJSON writes the given message as JSON, optionally with the
-// body included to a stream.
-func WriteMessageJSON(out io.Writer, message rabtap.TapMessage) error {
-	metadata, err := json.MarshalIndent(NewRabtapPersistentMessage(message), "", "  ")
+// marshalFunc marshals messages prior to writing them e.g. in JSON format
+type marshalFunc func(m interface{}) ([]byte, error)
+
+// JSONMarshmarshallIndent the given message as a formatted JSON
+func JSONMarshalIndent(m interface{}) ([]byte, error) {
+	return json.MarshalIndent(m, "", "  ")
+}
+
+// JSONMarshal marshalls the given message as a single line JSON
+func JSONMarshal(m interface{}) ([]byte, error) {
+	return json.Marshal(m)
+}
+
+// WriteMessage writes the given message using the proviced marshaller and writer
+func WriteMessage(out io.Writer, message rabtap.TapMessage, marshaller marshalFunc) error {
+	data, err := marshaller(NewRabtapPersistentMessage(message))
 	if err != nil {
 		return err
 	}
-	_, err = out.Write(metadata)
+	nl := []byte("\n")
+	_, err = out.Write(append(data, nl...))
 	return err
 }
 
@@ -120,14 +134,14 @@ func saveMessageBodyAsBlobFile(filename string, body []byte) error {
 	return writer.Flush()
 }
 
-func saveMessageAsJSONFile(filename string, message rabtap.TapMessage) error {
+func saveMessageAsJSONFile(filename string, message rabtap.TapMessage, marshaller marshalFunc) error {
 	file, err := os.Create(filename)
 	if err != nil {
 		return err
 	}
 	defer file.Close()
 	writer := bufio.NewWriter(file)
-	err = WriteMessageJSON(writer, message)
+	err = WriteMessage(writer, message, marshaller)
 	if err != nil {
 		return err
 	}
@@ -135,27 +149,26 @@ func saveMessageAsJSONFile(filename string, message rabtap.TapMessage) error {
 }
 
 // SaveMessageToRawFile writes a message to 2 files, one with the metadata, and
-// one with the payload
-func SaveMessageToRawFile(basename string, message rabtap.TapMessage) error {
+// one with the payload. The metadata will be serialized using the proviced marshaller
+func SaveMessageToRawFiles(basename string, message rabtap.TapMessage, marshaller marshalFunc) error {
 	filenameRaw := basename + ".dat"
-	filenameJSON := basename + ".json"
-	log.Debugf("saving message  %s (RAW) with meta data in %s", filenameRaw, filenameJSON)
+	filenameMeta := basename + ".json"
+	log.Debugf("saving message  %s (RAW) with meta data in %s", filenameRaw, filenameMeta)
 	err := saveMessageBodyAsBlobFile(filenameRaw, message.AmqpMessage.Body)
 	if err != nil {
 		return err
 	}
-	//return saveMessageAsJSONFile(filenameJSON, false, message)
 	// save metadata file without the body
 	oldBody := message.AmqpMessage.Body
 	message.AmqpMessage.Body = []byte{}
-	err = saveMessageAsJSONFile(filenameJSON, message)
+	err = saveMessageAsJSONFile(filenameMeta, message, marshaller)
 	message.AmqpMessage.Body = oldBody
 	return err
 }
 
 // SaveMessageToJSONFile writes a message to a single JSON file, where
 // the body will be BASE64 encoded
-func SaveMessageToJSONFile(filename string, message rabtap.TapMessage) error {
+func SaveMessageToJSONFile(filename string, message rabtap.TapMessage, marshaller marshalFunc) error {
 	log.Debugf("saving message to %s (JSON)", filename)
-	return saveMessageAsJSONFile(filename, message)
+	return saveMessageAsJSONFile(filename, message, marshaller)
 }